updated file structure for balance
